refactor: reusable squad stats
